requirements.txt
lyzr==0.1.35
python-dotenv==1.0.1
streamlit==1.33.0